requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
supabase==1.0.3
pandas==2.1.0
matplotlib==3.7.2
networkx==3.1
openai==0.28.0
pydantic==2.3.0
pytest==7.4.0
httpx<0.24.0
python-multipart==0.0.6
numpy==1.25.2
scipy==1.11.2
